feat: divide tab and music controller components
